+++ pybop/models/lithium_ion/echem.py
@@ -72,7 +72,9 @@ class SPM(BaseModel):
 
         self._electrode_soh = pybamm.lithium_ion.electrode_soh
 
-    def _check_params(self, inputs=None, allow_infeasible_solutions=True):
+    def _check_params(
+        self, inputs=None, parameter_set=None, allow_infeasible_solutions=True
+    ):
         """
         Check compatibility of the model parameters.
 
@@ -88,6 +90,7 @@ class SPM(BaseModel):
         bool
             A boolean which signifies whether the parameters are compatible.
         """
+        parameter_set = parameter_set or self._parameter_set
 
         electrode_params = [
             (
@@ -101,9 +104,7 @@ class SPM(BaseModel):
         ]
 
         related_parameters = {
-            key: inputs.get(key)
-            if inputs and key in inputs
-            else self._parameter_set[key]
+            key: inputs.get(key) if inputs and key in inputs else parameter_set[key]
             for pair in electrode_params
             for key in pair
         }
@@ -192,7 +193,9 @@ class SPMe(BaseModel):
 
         self._electrode_soh = pybamm.lithium_ion.electrode_soh
 
-    def _check_params(self, inputs=None, allow_infeasible_solutions=True):
+    def _check_params(
+        self, inputs=None, parameter_set=None, allow_infeasible_solutions=True
+    ):
         """
         Check compatibility of the model parameters.
 
@@ -208,6 +211,7 @@ class SPMe(BaseModel):
         bool
             A boolean which signifies whether the parameters are compatible.
         """
+        parameter_set = parameter_set or self._parameter_set
 
         electrode_params = [
             (
@@ -221,9 +225,7 @@ class SPMe(BaseModel):
         ]
 
         related_parameters = {
-            key: inputs.get(key)
-            if inputs and key in inputs
-            else self._parameter_set[key]
+            key: inputs.get(key) if inputs and key in inputs else parameter_set[key]
             for pair in electrode_params
             for key in pair
         }