+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Features
 
+- [#118](https://github.com/pybop-team/PyBOP/issues/118) - Adds example jupyter notebooks.
+- [#151](https://github.com/pybop-team/PyBOP/issues/151) - Adds a standalone version of the Problem class.
+
 ## Bug Fixes
 
 - [#63](https://github.com/pybop-team/PyBOP/issues/63) - Removes NLOpt Optimiser from future releases. This is to support deployment to the Apple M-Series platform.